remove old btree tests
